trying to upload local project to git

+++ drivers/fsl_tpm.h
@@ -87,7 +87,7 @@ typedef struct _tpm_chnl_pwm_signal_param
                                        In combined mode (available in some SoC's, this represents the
                                        channel pair number */
     tpm_pwm_level_select_t level; /*!< PWM output active level select */
-    float dutyCyclePercent;     /*!< PWM pulse width, value should be between 0 to 100
+    uint8_t dutyCyclePercent;     /*!< PWM pulse width, value should be between 0 to 100
                                        0=inactive signal(0% duty cycle)...
                                        100=always active signal (100% duty cycle)*/
 #if defined(FSL_FEATURE_TPM_HAS_COMBINE) && FSL_FEATURE_TPM_HAS_COMBINE
@@ -380,7 +380,7 @@ status_t TPM_SetupPwm(TPM_Type *base,
 void TPM_UpdatePwmDutycycle(TPM_Type *base,
                             tpm_chnl_t chnlNumber,
                             tpm_pwm_mode_t currentPwmMode,
-                            float dutyCyclePercent);
+                            uint8_t dutyCyclePercent);
 
 /*!
  * @brief Update the edge level selection for a channel
